Move statusenum to separate file to de-duplicate code
